requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langchain==0.3.2
langchain-community==0.3.1
langgraph==0.2.34
pyodbc==5.1.0
python-dotenv==1.0.1
coloredlogs==15.0.1
langchain-openai>=0.1.0